Memisahkan menu menjadi tagmenu dan todomenu, lalu menambahkan interface dan package helper
