+++ src/menu/IListing.java
@@ -1,0 +1,5 @@
+package menu;
+
+public interface IListing {
+	public void printMenuList();
+}
